feat: Move confirm delete dialog to a seperated component
