+++ apps/frontend/src/ui/inventory/edit/editable-inventory.component.ts
@@ -34,8 +34,7 @@ import {
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { EditableItemComponent } from './item/editable-item.component';
 import { debounceTime, distinctUntilChanged } from 'rxjs';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { Inject } from '@angular/core';
+import { ConfirmDeleteDialogComponent } from './confirm-delete-dialog.component';
 import { isSubset } from '@equip-track/shared';
 
 const formDuplicateValidator: ValidatorFn = (formArray: AbstractControl) => {
@@ -254,52 +253,4 @@ export class EditableInventoryComponent {
     }
     return null;
   };
-}
-
-@Component({
-  selector: 'app-confirm-delete-dialog',
-  standalone: true,
-  imports: [CommonModule, MatDialogModule, MatButtonModule, TranslateModule],
-  template: `
-    <h2 mat-dialog-title>{{ data.title }}</h2>
-    <mat-dialog-content>
-      <p>{{ data.message }}</p>
-    </mat-dialog-content>
-    <mat-dialog-actions align="end">
-      <button mat-button (click)="onCancel()">
-        {{ data.cancelText }}
-      </button>
-      <button mat-raised-button color="warn" (click)="onConfirm()">
-        {{ data.confirmText }}
-      </button>
-    </mat-dialog-actions>
-  `,
-  styles: [
-    `
-      mat-dialog-content {
-        min-width: 300px;
-        padding: 20px 0;
-      }
-    `,
-  ],
-})
-export class ConfirmDeleteDialogComponent {
-  constructor(
-    public dialogRef: MatDialogRef<ConfirmDeleteDialogComponent>,
-    @Inject(MAT_DIALOG_DATA)
-    public data: {
-      title: string;
-      message: string;
-      confirmText: string;
-      cancelText: string;
-    }
-  ) {}
-
-  onCancel(): void {
-    this.dialogRef.close(false);
-  }
-
-  onConfirm(): void {
-    this.dialogRef.close(true);
-  }
 }
